moved stdlib into ir. stdlib is now loaded automatically into context. added getNamespace to c interface
