+++ tests/bindings/python/test_namespace.py
@@ -5,7 +5,7 @@ import test_utils
 def test_get_name():
     context = coreir.Context()
     assert context.G.name == "_G"
-    stdlib = context.load_library("stdlib")
+    stdlib = context.get_namespace("stdlib")
     assert stdlib.name == "stdlib"
     add_instantiable = stdlib.instantiables["add"]
     assert add_instantiable.name == "add"
