+++ tests/cgra/Makefile
@@ -20,7 +20,7 @@ endif
 HOME = ../..
 INCS = -I$(HOME)/include -I.
 LPATH = -L$(HOME)/lib
-LIBS =  -Wl,-rpath,$(HOME)/lib -lcoreir-cgralib -lcoreir-stdlib -lcoreir-passes -lcoreir
+LIBS =  -Wl,-rpath,$(HOME)/lib -lcoreir-cgralib -lcoreir-passes -lcoreir
 SRCFILES = $(wildcard [^_]*.cpp)
 OBJS = $(patsubst %.cpp,build/%.o,$(SRCFILES))
 EXES = $(patsubst %.cpp,build/%,$(SRCFILES))
