+++ tests/unit/dadd4.cpp
@@ -1,5 +1,4 @@
 #include "coreir.h"
-#include "coreir-lib/stdlib.h"
 
 using namespace CoreIR;
 
@@ -11,7 +10,7 @@ int main() {
   
   Namespace* g = c->getGlobal();
   
-  Namespace* stdlib = CoreIRLoadLibrary_stdlib(c);
+  Namespace* stdlib = c->getNamespace("stdlib");
   
   //Declare add2 Generator
   Generator* add2 = stdlib->getGenerator("add");
